server/requirements.txt
flask
python-chess
gunicorn
flask-cors